# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Srulik's lab API
  version: 0.1.0
  description: >
    HTTP API for the Srulik's lab backend. This spec is the single source of truth for
    the generated API reference.
servers:
  - url: /
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      additionalProperties: true
      properties:
        error: { type: string }
        code: { type: string }
...
    ReportTodayResponse:
      type: object
      additionalProperties: true
    TranslateRequest:
      type: object
      required: [report, lang]
      properties:
        report: { type: object, additionalProperties: true }
        lang: { type: string, description: "Target language code" }
    TranslateResponse:
      type: object
      required: [report]
      properties:
        report: { type: object, nullable: true, additionalProperties: true }
    VideoDownloadUrlRequest:
      type: object
      required: [url]
      properties:
        url: { type: string }
    VideoLocalFileRequest:
      type: object
      required: [relativePath]
      properties:
        relativePath: { type: string }
...
    MailProductFlags:
      type: object
      required: [report, naftali, education, platform]
      properties:
        report: { type: boolean }
        naftali: { type: boolean }
        education: { type: boolean }
        platform: { type: boolean }
    MailPreferences:
      type: object
      required: [email, language, products]
      properties:
        email: { type: string }
        language: { type: string, enum: [en, he, ru] }
        products: { $ref: "#/components/schemas/MailProductFlags" }
    MailPreferencesUpdate:
      type: object
      properties:
        email: { type: string }
        language: { type: string, enum: [en, he, ru] }
        products:
          type: object
          additionalProperties: true
    MailSendDigestRequest:
      type: object
      properties:
        to:
          type: string
          description: "Optional; must match JWT email or saved mailing address"
    MailSendDigestResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        id: { type: string, nullable: true, description: "Resend message id when returned" }
...
paths:
...
  /api/report/today:
    get:
      tags: [Reports]
      summary: Fetch today's cached report (if any).
      parameters:
        - in: query
          name: scope
          required: false
          schema:
            type: string
            enum: [national, north]
          description: Report scope to fetch. Defaults to national.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReportTodayResponse" }
        "401":
          description: Unauthorized (when auth is required)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/video/download-url:
    post:
      tags: [Video]
      summary: Download a video from a URL and return its local path.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoDownloadUrlRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/video/local-file:
    post:
      tags: [Video]
      summary: Resolve a local video file path.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoLocalFileRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/municipalities:
    get:
      tags: [Municipalities]
      summary: Get municipality dashboard.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "502":
          description: Upstream failure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/geo/resolve:
    get:
      tags: [Geo]
      summary: Resolve a locality name to deterministic geo envelope (internal).
      description: |
        Returns a versioned geo envelope (`geo-envelope-2026-05-v3` when resolved).
        New resolves are **nested-only** (`resolution`, `classification`, `policy`, `audit`, `scopeDecision`, `matchEvidence`).
        `resolution.provenance` is one of `structured`, `text_inferred`, `message_level`, or `direct`.
        Text-inferred geo on news/radio/social sets `policy.usableForMetrics: false`.
        Optional query params `sourceType`, `provenance`, `reporterSubregionHint` are accepted by the service but not exposed here yet.
      parameters:
        - in: query
          name: name
          schema: { type: string }
          description: Locality name (Hebrew or English). Alias query param `q`.
        - in: query
          name: q
          schema: { type: string }
          description: Same as `name` when preferred.
      responses:
        "200":
          description: Geo envelope (resolved or unknown)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          description: Missing name
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Geo service unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/translate:
    post:
      tags: [Translation]
      summary: Translate an assessment to a target language (when enabled).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TranslateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranslateResponse" }
        "502":
          description: Translation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/mail/preferences:
    get:
      tags: [Mailing]
      summary: Get saved mailing preferences for the signed-in user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailPreferences" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [Mailing]
      summary: Update mailing preferences
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MailPreferencesUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailPreferences" }
        "400":
          description: Invalid email
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/mail/send-digest:
    post:
      tags: [Mailing]
      summary: Send one digest email now (Resend)
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MailSendDigestRequest" }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailSendDigestResponse" }
        "400":
          description: Bad request (no destination or invalid to)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: Provider error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Mailing not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...